requirements.txt
Django==5.2.1
djangorestframework==3.16.0
python-decouple==3.8
django-cors-headers==4.7.0
django-filter==25.1